Initial commit
This is the code for a desktop application written in Racket that
removes annoying tags from filenames. The code is almost ready for the
end user, though at current I'm having trouble making it run on OS X
(details in the README). This will probably compile and run just fine
on Windows, though I haven't checked recently.